Crypto.ECC

Description

Elliptic Curve Cryptography

Documentation

data Curve_P256R1

P256 Curve

also known as P256

Constructors

Curve_P256R1

data Curve_P384R1

Constructors

Curve_P384R1

data Curve_P521R1

Constructors

Curve_P521R1

data Curve_X25519

Constructors

Curve_X25519

data Curve_X448

Constructors

Curve_X448

data Curve_Edwards25519

Constructors

Curve_Edwards25519

class EllipticCurve curve where

Associated Types

type Point curve :: *

Point on an Elliptic Curve

type Scalar curve :: *

Scalar in the Elliptic Curve domain

Methods

curveGenerateScalar :: MonadRandom randomly => proxy curve -> randomly (Scalar curve)

Generate a new random scalar on the curve. The scalar will represent a number between 1 and the order of the curve non included

curveGenerateKeyPair :: MonadRandom randomly => proxy curve -> randomly (KeyPair curve)

Generate a new random keypair

curveSizeBits :: proxy curve -> Int

Get the curve size in bits

encodePoint :: ByteArray bs => proxy curve -> Point curve -> bs

Encode a elliptic curve point into binary form

decodePoint :: ByteArray bs => proxy curve -> bs -> CryptoFailable (Point curve)

Try to decode the binary form of an elliptic curve point

class EllipticCurve curve => EllipticCurveDH curve where

Minimal complete definition

ecdh

Methods

ecdhRaw :: proxy curve -> Scalar curve -> Point curve -> SharedSecret

Generate a Diffie hellman secret value.

This is generally just the .x coordinate of the resulting point, that is not hashed.

use pointSmul to keep the result in Point format.

WARNING: Curve implementations may return a special value or an exception when the public point lies in a subgroup of small order. This function is adequate when the scalar is in expected range and contributory behaviour is not needed. Otherwise use ecdh.

ecdh :: proxy curve -> Scalar curve -> Point curve -> CryptoFailable SharedSecret

Generate a Diffie hellman secret value and verify that the result is not the point at infinity.

This additional test avoids risks existing with function ecdhRaw. Implementations always return a CryptoError instead of a special value or an exception.

class EllipticCurve curve => EllipticCurveArith curve where

Methods

pointAdd :: proxy curve -> Point curve -> Point curve -> Point curve

Add points on a curve

pointNegate :: proxy curve -> Point curve -> Point curve

Negate a curve point

pointSmul :: proxy curve -> Scalar curve -> Point curve -> Point curve

Scalar Multiplication on a curve

data KeyPair curve

An elliptic curve key pair composed of the private part (a scalar), and the associated point.

Constructors

KeyPair

Fields

keypairGetPublic :: !(Point curve)

keypairGetPrivate :: !(Scalar curve)

newtype SharedSecret

Constructors

SharedSecret ScrubbedBytes