Crypto.Number.Basic

Description

Documentation

sqrti :: Integer -> (Integer, Integer)

sqrti returns two integers (l,b) so that l <= sqrt i <= b. The implementation is quite naive, use an approximation for the first number and use a dichotomy algorithm to compute the bound relatively efficiently.

gcde :: Integer -> Integer -> (Integer, Integer, Integer)

Get the extended GCD of two integer using integer divMod

gcde a b find (x,y,gcd(a,b)) where ax + by = d

areEven :: [Integer] -> Bool

Check if a list of integer are all even

log2 :: Integer -> Int

Compute the binary logarithm of a integer

numBits :: Integer -> Int

Compute the number of bits for an integer

numBytes :: Integer -> Int

Compute the number of bytes for an integer