Crypto.PubKey.Curve25519

Description

Curve25519 support

Documentation

data SecretKey

A Curve25519 Secret key

data PublicKey

A Curve25519 public key

data DhSecret

A Curve25519 Diffie Hellman secret related to a public key and a secret key.

Smart constructors

dhSecret :: ByteArrayAccess b => b -> CryptoFailable DhSecret

Create a DhSecret from a bytearray object

publicKey :: ByteArrayAccess bs => bs -> CryptoFailable PublicKey

Try to build a public key from a bytearray

secretKey :: ByteArrayAccess bs => bs -> CryptoFailable SecretKey

Try to build a secret key from a bytearray

methods

dh :: PublicKey -> SecretKey -> DhSecret

Compute the Diffie Hellman secret from a public key and a secret key

toPublic :: SecretKey -> PublicKey

Create a public key from a secret key