Crypto.ConstructHash.MiyaguchiPreneel

Description

provide the hash function construction method from block cipher https://en.wikipedia.org/wiki/One-way_compression_function

Documentation

compute

Arguments

:: (ByteArrayAccess bin, BlockCipher cipher)

=> bin
input message

-> MiyaguchiPreneel cipher
output tag

Compute Miyaguchi-Preneel one way compress using the infered block cipher. Only safe when KEY-SIZE equals to BLOCK-SIZE.

Simple usage mp' msg :: MiyaguchiPreneel AES128

compute'

Arguments

:: (ByteArrayAccess bin, BlockCipher cipher)

=> (Bytes -> cipher)
key build function to compute Miyaguchi-Preneel. care about block-size and key-size

-> bin
input message

-> MiyaguchiPreneel cipher
output tag

Compute Miyaguchi-Preneel one way compress using the supplied block cipher.

data MiyaguchiPreneel a