Crypto.Hash.Algorithms

Description

Definitions of known hash algorithms

Documentation

class HashAlgorithm a

Class representing hashing algorithms.

The interface presented here is update in place and lowlevel. the Hash module takes care of hidding the mutable interface properly.

Minimal complete definition

hashBlockSize, hashDigestSize, hashInternalContextSize, hashInternalInit, hashInternalUpdate, hashInternalFinalize

hash algorithms

data Blake2s_224

Blake2s (224 bits) cryptographic hash algorithm

Constructors

Blake2s_224

data Blake2sp_224

Blake2sp (224 bits) cryptographic hash algorithm

Constructors

Blake2sp_224

data Blake2s_256

Blake2s (256 bits) cryptographic hash algorithm

Constructors

Blake2s_256

data Blake2sp_256

Blake2sp (256 bits) cryptographic hash algorithm

Constructors

Blake2sp_256

data Blake2b_512

Blake2b (512 bits) cryptographic hash algorithm

Constructors

Blake2b_512

data Blake2bp_512

Blake2bp (512 bits) cryptographic hash algorithm

Constructors

Blake2bp_512

data MD2

MD2 cryptographic hash algorithm

Constructors

MD2

data MD4

MD4 cryptographic hash algorithm

Constructors

MD4

data MD5

MD5 cryptographic hash algorithm

Constructors

MD5

data SHA1

SHA1 cryptographic hash algorithm

Constructors

SHA1

data SHA224

SHA224 cryptographic hash algorithm

Constructors

SHA224

data SHA256

SHA256 cryptographic hash algorithm

Constructors

SHA256

data SHA384

SHA384 cryptographic hash algorithm

Constructors

SHA384

data SHA512

SHA512 cryptographic hash algorithm

Constructors

SHA512

data SHA512t_224

SHA512t (224 bits) cryptographic hash algorithm

Constructors

SHA512t_224

data SHA512t_256

SHA512t (256 bits) cryptographic hash algorithm

Constructors

SHA512t_256

data RIPEMD160

RIPEMD160 cryptographic hash algorithm

Constructors

RIPEMD160

data Tiger

Tiger cryptographic hash algorithm

Constructors

Tiger

data Keccak_224

Keccak (224 bits) cryptographic hash algorithm

Constructors

Keccak_224

data Keccak_256

Keccak (256 bits) cryptographic hash algorithm

Constructors

Keccak_256

data Keccak_384

Keccak (384 bits) cryptographic hash algorithm

Constructors

Keccak_384

data Keccak_512

Keccak (512 bits) cryptographic hash algorithm

Constructors

Keccak_512

data SHA3_224

SHA3 (224 bits) cryptographic hash algorithm

Constructors

SHA3_224

data SHA3_256

SHA3 (256 bits) cryptographic hash algorithm

Constructors

SHA3_256

data SHA3_384

SHA3 (384 bits) cryptographic hash algorithm

Constructors

SHA3_384

data SHA3_512

SHA3 (512 bits) cryptographic hash algorithm

Constructors

SHA3_512

data Skein256_224

Skein256 (224 bits) cryptographic hash algorithm

Constructors

Skein256_224

data Skein256_256

Skein256 (256 bits) cryptographic hash algorithm

Constructors

Skein256_256

data Skein512_224

Skein512 (224 bits) cryptographic hash algorithm

Constructors

Skein512_224

data Skein512_256

Skein512 (256 bits) cryptographic hash algorithm

Constructors

Skein512_256

data Skein512_384

Skein512 (384 bits) cryptographic hash algorithm

Constructors

Skein512_384

data Skein512_512

Skein512 (512 bits) cryptographic hash algorithm

Constructors

Skein512_512

data Whirlpool

Whirlpool cryptographic hash algorithm

Constructors

Whirlpool